+++ litemultiagent/agents/BaseAgent.py
@@ -1,26 +1,52 @@
-from typing import Any, Optional
+import logging
+from typing import List, Dict, Any, Optional
+from openai import OpenAI
 import json
 from concurrent.futures import ThreadPoolExecutor, as_completed
-
-from litemultiagent.core.config import AGENT_TO_MODEL, MODEL_COST, AgentLogger
-from litemultiagent.core.env import supabase
-
+from config import agent_to_model, model_cost
+from supabase import create_client, Client
+import os
+from dotenv import load_dotenv
+_ = load_dotenv()
 from litellm import completion
-from openai.types.chat import ChatCompletion
+from datetime import datetime
+import csv
+import json
+import os
+from datetime import datetime
+
+import csv
+import json
+import os
+from datetime import datetime
+
+logger = logging.getLogger(__name__)
+
+# Initialize Supabase client only if environment variables are set
+url = os.getenv("SUPABASE_URL")
+key = os.getenv("SUPABASE_ANON_KEY")
+supabase: Optional[Client] = None
+if url and key:
+    try:
+        supabase = create_client(url, key)
+    except Exception as e:
+        logger.error(f"Failed to initialize Supabase client: {e}")
+
 
 class Agent:
-    def __init__(self, agent_name: str, tools: list[dict[str, Any]], available_tools: dict[str, callable],
-                 meta_task_id: Optional[str] = None, task_id: Optional[int] = None):
+    def __init__(self, agent_name: str, tools: List[Dict[str, Any]], available_tools: Dict[str, callable],
+                 meta_task_id: Optional[str] = None, task_id: Optional[int] = None, save_to="csv", log="log"):
         self.agent_name = agent_name
         self.tools = tools
         self.available_tools = available_tools
-        self.model_name = AGENT_TO_MODEL[agent_name]["model_name"]
-        self.tool_choice = AGENT_TO_MODEL[agent_name]["tool_choice"]
+        self.model_name = agent_to_model[agent_name]["model_name"]
+        self.tool_choice = agent_to_model[agent_name]["tool_choice"]
         self.messages = []
         self.meta_task_id = meta_task_id
         self.task_id = task_id
+        self.save_to = save_to
+        self.log = log
 
-        
 
 
     def send_prompt(self, content: str) -> str:
@@ -38,7 +64,10 @@ class Agent:
                 messages=self.messages
             )
             self._log_response(response, depth)
-            self._save_to_supabase(response, depth)
+            if self.save_to == "supabase":
+                self._save_to_supabase(response, depth)
+            if self.save_to == "csv":
+                self._save_to_csv(response, depth)
             message = response.choices[0].message
             self.messages.append(message)
             return message.content
@@ -50,8 +79,12 @@ class Agent:
             tool_choice=self.tool_choice
         )
 
+        # whether it has tool call or not
         self._log_response(response, depth)
-        self._save_to_supabase(response, depth)
+        if self.save_to == "supabase":
+            self._save_to_supabase(response, depth)
+        if self.save_to == "csv":
+            self._save_to_csv(response, depth)
 
         tool_calls = response.choices[0].message.tool_calls
 
@@ -69,12 +102,13 @@ class Agent:
         self.messages.append(tool_call_message)
         tool_responses = self._process_tool_calls(tool_calls)
         self.messages.extend(tool_responses)
+        import pdb; pdb.set_trace()
 
         return self._send_completion_request(depth + 1)
 
-    def _process_tool_calls(self, tool_calls: list[dict[str, Any]]) -> list[dict[str, Any]]:
+    def _process_tool_calls(self, tool_calls: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
         tool_call_responses = []
-        AgentLogger.info(f"Number of function calls: {len(tool_calls)}")
+        logger.info(f"Number of function calls: {len(tool_calls)}")
 
         with ThreadPoolExecutor(max_workers=None) as executor:
             future_to_tool_call = {executor.submit(self._process_single_tool_call, tool_call): tool_call for tool_call
@@ -87,7 +121,7 @@ class Agent:
 
         return tool_call_responses
 
-    def _process_single_tool_call(self, tool_call: dict[str, Any]) -> dict[str, Any]:
+    def _process_single_tool_call(self, tool_call: Dict[str, Any]) -> Dict[str, Any]:
         tool_call_id = tool_call["id"]
         function_name = tool_call["function"]["name"]
         function_args = json.loads(tool_call["function"]["arguments"])
@@ -96,27 +130,69 @@ class Agent:
 
         try:
             function_response = function_to_call(**function_args)
-            AgentLogger.info(f'Function name: {function_name}, function args: {function_args}')
+            logger.info(f'Function name: {function_name}, function args: {function_args}')
             tool_response_message = {
                 "tool_call_id": tool_call_id,
                 "role": "tool",
                 "name": function_name,
                 "content": str(function_response)
             }
-            AgentLogger.info(f'Function name: {function_name}, function response: {str(function_response)}')
+            logger.info(f'Function name: {function_name}, function response: {str(function_response)}')
             return tool_response_message
         except Exception as e:
-            AgentLogger.error(f"Error while calling function <{function_name}>: {e}")
+            logger.error(f"Error while calling function <{function_name}>: {e}")
             return None
 
     def _log_response(self, response, depth):
-        AgentLogger.info(
+        logger.info(
             f'Agent: {self.agent_name}, prompt tokens: {response.usage.prompt_tokens}, completion tokens: {response.usage.completion_tokens}')
-        AgentLogger.info(f'Agent: {self.agent_name}, depth: {depth}, response: {response}')
+        logger.info(f'Agent: {self.agent_name}, depth: {depth}, response: {response}')
 
-    def _save_to_supabase(self, response: ChatCompletion, depth):
+
+    def _save_to_csv(self, response, depth):
+        usage_dict = self._extract_cost(response)
+        data = {
+            "meta_task_id": self.meta_task_id,
+            "task_id": self.task_id,
+            "agent": self.agent_name,
+            "depth": depth,
+            "role": "assistant",
+            "response": json.dumps(response.choices[0].message.model_dump()),
+            "prompt_tokens": response.usage.prompt_tokens,
+            "completion_tokens": response.usage.completion_tokens,
+            "input_cost": usage_dict["input_cost"],
+            "output_cost": usage_dict["output_cost"],
+            "total_cost": usage_dict["total_cost"],
+            "model_name": self.model_name,
+            "timestamp": datetime.now().isoformat()
+        }
+
+        filename = os.path.join(self.log, f"multiagent_data_{datetime.now().strftime('%Y%m%d')}.csv")
+        file_exists = os.path.isfile(filename)
+
+        # Ensure the directory exists
+        os.makedirs(os.path.dirname(filename), exist_ok=True)
+
+        # If file doesn't exist, create it with header
+        if not file_exists:
+            with open(filename, 'w', newline='') as csvfile:
+                fieldnames = list(data.keys())
+                writer = csv.DictWriter(csvfile, fieldnames=fieldnames)
+                writer.writeheader()
+            logger.info(f"Created new CSV file with header: {filename}")
+
+        # Append data to the file
+        with open(filename, 'a', newline='') as csvfile:
+            fieldnames = list(data.keys())
+            writer = csv.DictWriter(csvfile, fieldnames=fieldnames)
+            writer.writerow(data)
+
+        logger.info(f"Data saved to CSV: {filename}")
+
+
+    def _save_to_supabase(self, response, depth):
         if supabase is None:
-            AgentLogger.warning("Supabase client is not initialized. Skipping database save.")
+            logger.warning("Supabase client is not initialized. Skipping database save.")
             return
 
         usage_dict = self._extract_cost(response)
@@ -138,16 +214,16 @@ class Agent:
         try:
             supabase.table("multiagent").insert(data).execute()
         except Exception as e:
-           AgentLogger.error(f"Failed to save data to Supabase: {e}")
+            logger.error(f"Failed to save data to Supabase: {e}")
 
-    def _extract_cost(self, response: ChatCompletion):
+    def _extract_cost(self, response):
         prompt_tokens = response.usage.prompt_tokens
         completion_tokens = response.usage.completion_tokens
         total_tokens = response.usage.total_tokens
 
         # Pricing for gpt-4-0314
-        input_price_per_1m = MODEL_COST[self.model_name]["input_price_per_1m"]
-        output_price_per_1m = MODEL_COST[self.model_name]["output_price_per_1m"]
+        input_price_per_1m = model_cost[self.model_name]["input_price_per_1m"]
+        output_price_per_1m = model_cost[self.model_name]["output_price_per_1m"]
 
         input_cost = (prompt_tokens / 1000) * input_price_per_1m
         output_cost = (completion_tokens / 1000) * output_price_per_1m
