+++ litemultiagent/requirements.txt
@@ -1,0 +1,12 @@
+openai==1.34.0
+python-dotenv==1.0.1
+pydantic==2.7.4
+requests
+supabase
+langchain==0.2.7
+bs4==0.0.2
+chromadb==0.4.14
+tiktoken==0.7.0
+pypdf==4.3.1
+litellm==1.42.1
+langchain-community==0.2.7
